Application structure and code remodeled
- separate User,MonthlyCosts and MonthlyCostsSummary controllers;
-  changes to repositories;
- MonthlyCostsSummaryNotFoundException;
- CustomizedResponseEntityExceptionHandler - method added - handleMonthlyCostsSummaryNotFound;
- changes to entities - createDate,deleteDate and lastModifiedDate added, data relation changes;
- MonthlyCosts entity - id type change - to Long;
- controllers and services - changes to methods

+++ src/main/java/com/jerzymaj/budgetmanagement/budget_management_app/jpa_repositories/MonthlyCostsSummaryRepository.java
@@ -1,11 +1,12 @@
 package com.jerzymaj.budgetmanagement.budget_management_app.jpa_repositories;
 
 import com.jerzymaj.budgetmanagement.budget_management_app.models.MonthlyCostsSummary;
+import io.swagger.v3.oas.models.security.SecurityScheme;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
 public interface MonthlyCostsSummaryRepository extends JpaRepository<MonthlyCostsSummary, Integer> {
 
-       Optional<MonthlyCostsSummary> findByMonthlyCostsId(int monthlyCostsIdd);
+       Optional<MonthlyCostsSummary> findByMonthlyCostsId(Long monthlyCostsIdd);
 }
